refactor(utils): move image utilities to new package `filters/imgutil`
