+++ filters/binarization/concurrent.go
@@ -4,7 +4,7 @@ import (
 	"image"
 	"sync"
 
-	"dougdomingos.com/image-filters/filters/partitioning"
+	"dougdomingos.com/image-filters/filters/imgutil"
 	"dougdomingos.com/image-filters/utils"
 )
 
@@ -14,8 +14,8 @@ import (
 func concurrentBinarization(img *image.RGBA) {
 	var (
 		bounds      = img.Bounds()
-		numWorkers  = partitioning.GetNumberOfWorkers(bounds)
-		imageStrips = partitioning.GetVerticalPartitions(bounds, numWorkers)
+		numWorkers  = imgutil.GetNumberOfWorkers(bounds)
+		imageStrips = imgutil.GetVerticalPartitions(bounds, numWorkers)
 		threshold   = otsuThreshold(img, bounds)
 		wg          sync.WaitGroup
 	)
